requirements.txt
anthropic>=0.25.0           # Claude用（任意）
groq>=0.9.0                 # Groq用（任意）
google-genai>=1.0.0         # Gemini用（新SDK）
faster-whisper>=1.0.0
httpx>=0.27.0
sounddevice>=0.4.6
soundfile>=0.12.1
websockets>=12.0
pyyaml>=6.0
pydantic>=2.0
python-dotenv>=1.0.0
numpy>=1.24.0
